+++ core-ng/src/main/java/core/framework/impl/resource/PoolMetrics.java
@@ -1,0 +1,26 @@
+package core.framework.impl.resource;
+
+import core.framework.impl.log.stat.Metrics;
+
+import java.util.Map;
+
+/**
+ * @author neo
+ */
+public class PoolMetrics implements Metrics {
+    private final Pool<?> pool;
+
+    public PoolMetrics(Pool<?> pool) {
+        this.pool = pool;
+    }
+
+    @Override
+    public void collect(Map<String, Double> stats) {
+        stats.put(statName("total_count"), (double) pool.totalCount());
+        stats.put(statName("idle_count"), (double) pool.idleCount());
+    }
+
+    String statName(String statName) {
+        return "pool_" + pool.name + '_' + statName;
+    }
+}