+++ core-ng/src/main/java/core/framework/module/CacheConfig.java
@@ -8,6 +8,7 @@ import core.framework.impl.cache.LocalCacheStore;
 import core.framework.impl.cache.RedisCacheStore;
 import core.framework.impl.module.ModuleContext;
 import core.framework.impl.redis.RedisImpl;
+import core.framework.impl.resource.PoolMetrics;
 import core.framework.impl.web.management.CacheController;
 import core.framework.util.ASCII;
 import core.framework.util.Exceptions;
@@ -76,6 +77,7 @@ public final class CacheConfig {
             redis.timeout(Duration.ofSeconds(1));   // for cache, use shorter timeout than default redis config
             context.shutdownHook.add(redis::close);
             context.backgroundTask().scheduleWithFixedDelay(redis.pool::refresh, Duration.ofMinutes(5));
+            context.metrics.add(new PoolMetrics(redis.pool));
             configureCacheManager(new RedisCacheStore(redis));
         }
     }