+++ core-ng/src/test/java/core/framework/impl/resource/TestPoolResource.java
@@ -1,0 +1,13 @@
+package core.framework.impl.resource;
+
+/**
+ * @author neo
+ */
+class TestPoolResource implements AutoCloseable {
+    boolean closed;
+
+    @Override
+    public void close() {
+        closed = true;
+    }
+}