+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * bean: removed "javax.inject:javax.inject:1", replaced with core.framework.inject, removed constructor injection support (to simplify and prepare for JDK 9) 
 * jdk: make built target compatible with JDK 9, as first step of java 9 migration (some of toolchain does not support java 9 yet) 
 * kafka: update to 1.0.0
+* pool: monitor pool size for both db and redis/cache/session
 
 ### 5.1.1 (10/17/2017 - 10/27/2017)
 * pool: refactor and simplify resource pool